feat: add to/from Basis points to PercentageUtils

+++ packages/core-contracts/test/mocks/ConfigurationManagerMock.sol
@@ -12,7 +12,7 @@ abstract contract ConfigurationManagerMock is IConfigurationManager {
 
     function raft() external pure returns (address) {}
 
-    function tipRate() external pure returns (uint8) {}
+    function tipRate() external pure returns (uint256) {}
 
     function setRaft(address) external pure {}
